clean up react app, make more OOP
